requirements.txt
google-generativeai==0.3.2
python-dotenv==1.0.0
colorama==0.4.6
flask==2.3.3
flask-socketio==5.3.6 